Add sql results interface
